+++ R/lasground.r
@@ -29,7 +29,7 @@
 #' Classify points as ground or not ground
 #'
 #' Implements a Progressive Morphological Filter for segmentation of ground points.
-#' The function updates the field \code{Classification} of the input LAS obect. The points
+#' The function updates the field \code{Classification} of the input LAS object. The points
 #' classified as 'ground' are assigned a value of 2 according to las specifications (See the ASPRS
 #' documentation for the
 #' \href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}).
@@ -64,10 +64,12 @@ lasground = function(.las, MaxWinSize = 20, Slope = 1.0, InitDist = 0.5, MaxDist
 {
   . <- X <- Y <- Z <- Classification <- NULL
 
+  stopifnotlas(.las)
+
   dots = list(...)
 
-  if(is.null(dots$base)) dots$base = 2
-  if(is.null(dots$exponential)) dots$exponential = TRUE
+  if (is.null(dots$base)) dots$base = 2
+  if (is.null(dots$exponential)) dots$exponential = TRUE
 
   exponential = dots$exponential
   base = dots$base
@@ -75,11 +77,11 @@ lasground = function(.las, MaxWinSize = 20, Slope = 1.0, InitDist = 0.5, MaxDist
   cloud = .las@data[, .(X,Y,Z)]
   cloud[, idx := 1:dim(cloud)[1]]
 
-  if("Classification" %in% names(.las@data))
+  if ("Classification" %in% names(.las@data))
   {
     nground = fast_countequal(.las@data$Classification, 2)
 
-    if(nground > 0)
+    if (nground > 0)
     {
       warning(paste0("Orginal dataset already contains ", nground, " ground points. These points were reclassified as 'unclassified' before to perform a new ground classification."), call. = FALSE)
       .las@data[Classification == 2, Classification := 0]
@@ -89,6 +91,8 @@ lasground = function(.las, MaxWinSize = 20, Slope = 1.0, InitDist = 0.5, MaxDist
   {
     .las@data[, Classification := 0]
   }
+
+  verbose("Progressive morphological filter...")
 
   idx = ProgressiveMorphologicalFilter(cloud, MaxWinSize, Slope, InitDist, MaxDist, CellSize, base, exponential)
 
@@ -112,9 +116,9 @@ ProgressiveMorphologicalFilter = function(cloud, MaxWinSize, Slope, InitDist, Ma
   {
     # Determine the initial window size.
     if (exponential)
-      window_size = CellSize * ( (2.0 * base^iteration) + 1 )
+      window_size = CellSize * ((2.0 * base^iteration) + 1 )
     else
-      window_size = CellSize * (2.0 * (iteration+1) * base + 1)
+      window_size = CellSize * (2.0 * (iteration + 1) * base + 1)
 
     # Calculate the height threshold to be used in the next iteration.
     if (iteration == 0)
@@ -135,7 +139,10 @@ ProgressiveMorphologicalFilter = function(cloud, MaxWinSize, Slope, InitDist, Ma
   # Filter ground returns using a progressive morphological filter
   for (i in 1:length(window_sizes))
   {
-    Z_f = MorphologicalOpening(cloud$X, cloud$Y, cloud$Z, window_sizes[i])
+    verbose(paste0("Pass ", i, " of ", length(window_sizes), "..."))
+    verbose(paste0("Windows size = ", window_sizes[i], " ; height_threshold = ", height_thresholds[i]))
+
+    Z_f = MorphologicalOpening(cloud$X, cloud$Y, cloud$Z, window_sizes[i], LIDROPTIONS("progress"))
 
     # Find indices of the points whose difference between the source and
     # filtered point clouds is less than the current height threshold