+++ src/RcppExports.cpp
@@ -33,17 +33,19 @@ BEGIN_RCPP
 END_RCPP
 }
 // algo_li2012
-IntegerVector algo_li2012(NumericVector X, NumericVector Y, NumericVector Z, NumericVector dt, double R);
-RcppExport SEXP lidR_algo_li2012(SEXP XSEXP, SEXP YSEXP, SEXP ZSEXP, SEXP dtSEXP, SEXP RSEXP) {
+IntegerVector algo_li2012(NumericVector X, NumericVector Y, const NumericVector Z, double dt1, double dt2, double R, bool displaybar);
+RcppExport SEXP lidR_algo_li2012(SEXP XSEXP, SEXP YSEXP, SEXP ZSEXP, SEXP dt1SEXP, SEXP dt2SEXP, SEXP RSEXP, SEXP displaybarSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type X(XSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type Y(YSEXP);
-    Rcpp::traits::input_parameter< NumericVector >::type Z(ZSEXP);
-    Rcpp::traits::input_parameter< NumericVector >::type dt(dtSEXP);
+    Rcpp::traits::input_parameter< const NumericVector >::type Z(ZSEXP);
+    Rcpp::traits::input_parameter< double >::type dt1(dt1SEXP);
+    Rcpp::traits::input_parameter< double >::type dt2(dt2SEXP);
     Rcpp::traits::input_parameter< double >::type R(RSEXP);
-    rcpp_result_gen = Rcpp::wrap(algo_li2012(X, Y, Z, dt, R));
+    Rcpp::traits::input_parameter< bool >::type displaybar(displaybarSEXP);
+    rcpp_result_gen = Rcpp::wrap(algo_li2012(X, Y, Z, dt1, dt2, R, displaybar));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -95,23 +97,9 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// get_normales
-NumericMatrix get_normales(IntegerMatrix M, NumericMatrix X, int size, bool edge_size);
-RcppExport SEXP lidR_get_normales(SEXP MSEXP, SEXP XSEXP, SEXP sizeSEXP, SEXP edge_sizeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< IntegerMatrix >::type M(MSEXP);
-    Rcpp::traits::input_parameter< NumericMatrix >::type X(XSEXP);
-    Rcpp::traits::input_parameter< int >::type size(sizeSEXP);
-    Rcpp::traits::input_parameter< bool >::type edge_size(edge_sizeSEXP);
-    rcpp_result_gen = Rcpp::wrap(get_normales(M, X, size, edge_size));
-    return rcpp_result_gen;
-END_RCPP
-}
 // MorphologicalOpening
-NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVector Z, double resolution);
-RcppExport SEXP lidR_MorphologicalOpening(SEXP XSEXP, SEXP YSEXP, SEXP ZSEXP, SEXP resolutionSEXP) {
+NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVector Z, double resolution, bool displaybar);
+RcppExport SEXP lidR_MorphologicalOpening(SEXP XSEXP, SEXP YSEXP, SEXP ZSEXP, SEXP resolutionSEXP, SEXP displaybarSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -119,7 +107,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericVector >::type Y(YSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type Z(ZSEXP);
     Rcpp::traits::input_parameter< double >::type resolution(resolutionSEXP);
-    rcpp_result_gen = Rcpp::wrap(MorphologicalOpening(X, Y, Z, resolution));
+    Rcpp::traits::input_parameter< bool >::type displaybar(displaybarSEXP);
+    rcpp_result_gen = Rcpp::wrap(MorphologicalOpening(X, Y, Z, resolution, displaybar));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -152,8 +141,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // points_in_polygons
-IntegerVector points_in_polygons(Rcpp::List vertx, Rcpp::List verty, NumericVector pointx, NumericVector pointy);
-RcppExport SEXP lidR_points_in_polygons(SEXP vertxSEXP, SEXP vertySEXP, SEXP pointxSEXP, SEXP pointySEXP) {
+IntegerVector points_in_polygons(Rcpp::List vertx, Rcpp::List verty, NumericVector pointx, NumericVector pointy, bool displaybar);
+RcppExport SEXP lidR_points_in_polygons(SEXP vertxSEXP, SEXP vertySEXP, SEXP pointxSEXP, SEXP pointySEXP, SEXP displaybarSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -161,7 +150,36 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::List >::type verty(vertySEXP);
     Rcpp::traits::input_parameter< NumericVector >::type pointx(pointxSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type pointy(pointySEXP);
-    rcpp_result_gen = Rcpp::wrap(points_in_polygons(vertx, verty, pointx, pointy));
+    Rcpp::traits::input_parameter< bool >::type displaybar(displaybarSEXP);
+    rcpp_result_gen = Rcpp::wrap(points_in_polygons(vertx, verty, pointx, pointy, displaybar));
+    return rcpp_result_gen;
+END_RCPP
+}
+// tinfo
+NumericMatrix tinfo(IntegerMatrix M, NumericMatrix X);
+RcppExport SEXP lidR_tinfo(SEXP MSEXP, SEXP XSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< IntegerMatrix >::type M(MSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type X(XSEXP);
+    rcpp_result_gen = Rcpp::wrap(tinfo(M, X));
+    return rcpp_result_gen;
+END_RCPP
+}
+// tsearch
+IntegerVector tsearch(NumericVector x, NumericVector y, IntegerMatrix elem, NumericVector xi, NumericVector yi, bool diplaybar);
+RcppExport SEXP lidR_tsearch(SEXP xSEXP, SEXP ySEXP, SEXP elemSEXP, SEXP xiSEXP, SEXP yiSEXP, SEXP diplaybarSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type y(ySEXP);
+    Rcpp::traits::input_parameter< IntegerMatrix >::type elem(elemSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type xi(xiSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type yi(yiSEXP);
+    Rcpp::traits::input_parameter< bool >::type diplaybar(diplaybarSEXP);
+    rcpp_result_gen = Rcpp::wrap(tsearch(x, y, elem, xi, yi, diplaybar));
     return rcpp_result_gen;
 END_RCPP
 }