+++ src/cxx_utils.cpp
@@ -86,81 +86,6 @@ int fast_countover(NumericVector x, double t)
   return n;
 }
 
-IntegerVector which_true(LogicalVector x)
-{
-  int n = 0;
-  IntegerVector y;
-
-  for (LogicalVector::iterator it = x.begin(), end = x.end() ; it != end ; ++it)
-  {
-    if (*it)
-      y.push_back(n);
-
-    n++;
-  }
-
-  return y;
-}
-
-int which_max(NumericVector x)
-{
-  int n = -1;
-  double max = -INFINITY;
-
-  for (int i = 0, end = x.length() ; i < end ; i++)
-  {
-    if (x(i) >= max)
-    {
-      max = x(i);
-      n = i;
-    }
-  }
-
-  return n;
-}
-
-// [[Rcpp::export]]
-NumericMatrix get_normales(IntegerMatrix M, NumericMatrix X, int size, bool edge_size = false)
-{
-  NumericMatrix N(size, 5);
-  std::fill(N.begin(), N.end(), NA_REAL);
-
-  for(int i = 0, end = M.nrow(); i < end ; i++)
-  {
-    int p1 = M(i,0)-1;
-    int p2 = M(i,1)-1;
-    int p3 = M(i,2)-1;
-    int j  = M(i,3)-1;
-
-    NumericVector A = NumericVector::create(X(p1,0), X(p1,1), X(p1,2));
-    NumericVector B = NumericVector::create(X(p2,0), X(p2,1), X(p2,2));
-    NumericVector C = NumericVector::create(X(p3,0), X(p3,1), X(p3,2));
-
-    NumericVector u = A - B;
-    NumericVector v = A - C;
-    NumericVector w = B - C;
-
-    NumericVector n = NumericVector::create(u(1)*v(2)-u(2)*v(1), u(2)*v(0)-u(0)*v(2), u(0)*v(1)-u(1)*v(0));
-    n.push_back(sum(-n*C));
-
-    N(j,0) = n(0);
-    N(j,1) = n(1);
-    N(j,2) = n(2);
-    N(j,3) = n(3);
-
-    if(edge_size)
-    {
-      u.erase(2);
-      v.erase(2);
-      w.erase(2);
-      NumericVector e = NumericVector::create(sqrt(sum(pow(u, 2))), sqrt(sum(pow(v, 2))), sqrt(sum(pow(w, 2))));
-      N(j,4) = max(e);
-    }
-  }
-
-  return N;
-}
-
 IntegerMatrix which_equal(IntegerMatrix mtx, double val)
 {
   int l = mtx.nrow();
@@ -199,9 +124,21 @@ NumericVector filter_xx(NumericMatrix x, IntegerMatrix y)
   return(out);
 }
 
-NumericVector distance(NumericVector x1, NumericVector y1, double x2, double y2)
+NumericVector sqdistance(NumericVector x1, NumericVector y1, double x2, double y2)
 {
-  NumericVector y = sqrt( pow((x1-x2),2) + pow((y1-y2),2));
+  int n = x1.length();
+  NumericVector y(n);
+  NumericVector::iterator i1, i2, i3, end1, end2, end3;
+
+  for( i1 = x1.begin(), i2 = y1.begin(), i3 = y.begin(), end1 = x1.end(), end2 = y1.end(), end3 = y.end();
+       i1 < end1 && i2 < end2 && i3 < end3;
+        ++i1, ++i2 , ++i3)
+  {
+    double dx = *i1-x2;
+    double dy = *i2-y2;
+    *i3 = dx * dx + dy * dy;
+  }
+
   return y;
 }