+++ src/morphology.cpp
@@ -27,7 +27,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
 ===============================================================================
 */
 
-
+// [[Rcpp::depends(RcppProgress)]]
+#include <progress.hpp>
 #include <Rcpp.h>
 #include <limits>
 #include "QuadTree.h"
@@ -35,7 +36,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>
 using namespace Rcpp;
 
 // [[Rcpp::export]]
-NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVector Z, double resolution)
+NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVector Z, double resolution, bool displaybar = false)
 {
   int n = X.length();
   double half_res = resolution / 2;
@@ -43,13 +44,20 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
   NumericVector Z_temp = clone(Z);
   NumericVector Z_out  = clone(Z);
 
-  QuadTree tree = QuadTree::create(as< std::vector<double> >(X),as< std::vector<double> >(Y));
+  QuadTree *tree = QuadTree::create(as< std::vector<double> >(X),as< std::vector<double> >(Y));
+
+  Progress p(2*n, displaybar);
 
   // Dilate
   for (int i = 0 ; i < n ; i++)
   {
+    if (Progress::check_abort() )
+      return Z_out;
+    else
+      p.update(i);
+
     std::vector<Point*> pts;
-    tree.rect_lookup(X[i], Y[i], half_res, pts);
+    tree->rect_lookup(X[i], Y[i], half_res, half_res, pts);
 
     double min_pt(std::numeric_limits<double>::max());
 
@@ -69,8 +77,13 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
   // erode
   for (int i = 0 ; i < n ; i++)
   {
+    if (Progress::check_abort() )
+      return Z_out;
+    else
+      p.update(i+n);
+
     std::vector<Point*> pts;
-    tree.rect_lookup(X[i], Y[i], half_res, pts);
+    tree->rect_lookup(X[i], Y[i], half_res, half_res, pts);
 
     double max_pt(std::numeric_limits<double>::min());
 
@@ -84,6 +97,8 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
 
     Z_out[i] = max_pt;
   }
+
+  delete tree;
 
   return Z_out;
 }